Somfy RFY  and vertical vs horizontal support
‼️BREAKING CHANGE‼️

Added support for vertical covers that have a different duration for close (=down) versus open (=up).
Somfy covers should work now, but this has not yet been tested (on my to do list, but probably not for the coming weeks, so give me some feedback :-)).

Previous configurations had a "time" parameter that is now replaced with "duration" for horizontal covers and "duration_up" and "duration_down" for vertical covers.